Move slow rdctl tests to a separate file.
