+++ e2e/rdctl.e2e.spec.ts
@@ -29,11 +29,10 @@ import yaml from 'yaml';
 
 import { NavPage } from './pages/nav-page';
 import {
-  createDefaultSettings, getAlternateSetting, kubectl, retry, startRancherDesktop, teardown, tool, waitForRestartVM,
+  createDefaultSettings, kubectl, retry, startRancherDesktop, teardown,
 } from './utils/TestUtils';
 
 import {
-  ContainerEngine,
   Settings,
   defaultSettings,
   CURRENT_SETTINGS_VERSION,
@@ -43,7 +42,6 @@ import {
   SecurityModel,
   VMType,
 } from '@pkg/config/settings';
-import { PathManagementStrategy } from '@pkg/integrations/pathManager';
 import { ServerState } from '@pkg/main/commandServer/httpCommandServer';
 import { spawnFile } from '@pkg/utils/childProcess';
 import paths from '@pkg/utils/paths';
@@ -136,7 +134,7 @@ test.describe('Command server', () => {
     const dataRaw = await fs.promises.readFile(dataPath, 'utf-8');
 
     serverState = JSON.parse(dataRaw);
-    expect(serverState).toEqual(expect.objectContaining({
+    expect(serverState).toEqual(expect.objectContaining( {
       user:     expect.any(String),
       password: expect.any(String),
       port:     expect.any(Number),
@@ -794,123 +792,6 @@ test.describe('Command server', () => {
         expect(stderr).not.toContain('Usage:');
       });
 
-      test.describe('settings v5 migration', () => {
-        /**
-         * Note issue https://github.com/rancher-sandbox/rancher-desktop/issues/3829
-         * calls for removing unrecognized fields in the existing settings.json file
-         * Currently we're ignoring unrecognized fields in the PUT payload -- to complain about
-         * them calls for another issue.
-         */
-        test('rejects old settings', async() => {
-          const oldSettings: Settings = JSON.parse((await rdctl(['list-settings'])).stdout);
-          const body: any = {
-            // type 'any' because as far as the current configuration code is concerned,
-            // it's an object with random fields and values
-            version:    CURRENT_SETTINGS_VERSION,
-            kubernetes: {
-              memoryInGB:      oldSettings.virtualMachine.memoryInGB + 1,
-              numberCPUs:      oldSettings.virtualMachine.numberCPUs + 1,
-              containerEngine: getAlternateSetting(oldSettings, 'containerEngine.name', ContainerEngine.CONTAINERD, ContainerEngine.MOBY),
-              suppressSudo:    oldSettings.application.adminAccess,
-            },
-            telemetry: !oldSettings.application.telemetry.enabled,
-            updater:   !oldSettings.application.updater.enabled,
-            debug:     !oldSettings.application.debug,
-          };
-          const addPathManagementStrategy = (oldSettings: Settings, body: any) => {
-            body.pathManagementStrategy = getAlternateSetting(oldSettings,
-              'application.pathManagementStrategy',
-              PathManagementStrategy.Manual,
-              PathManagementStrategy.RcFiles);
-          };
-
-          switch (os.platform()) {
-          case 'darwin':
-            body.kubernetes.experimental ??= {};
-            body.kubernetes.experimental.socketVMNet = !oldSettings.experimental.virtualMachine.socketVMNet;
-            addPathManagementStrategy(oldSettings, body);
-            break;
-          case 'linux':
-            addPathManagementStrategy(oldSettings, body);
-            break;
-          case 'win32':
-            body.kubernetes.WSLIntegrations ??= {};
-            body.kubernetes.WSLIntegrations.bosco = true;
-            body.kubernetes.hostResolver = !oldSettings.virtualMachine.hostResolver;
-          }
-          const { stdout, stderr, error } = await rdctl(['api', '/v1/settings', '-X', 'PUT', '-b', JSON.stringify(body)]);
-
-          expect({
-            stdout, stderr, error,
-          }).toEqual({
-            stdout: expect.stringContaining('no changes necessary'),
-            stderr: '',
-            error:  undefined,
-          });
-          const newSettings: Settings = JSON.parse((await rdctl(['list-settings'])).stdout);
-
-          expect(newSettings).toEqual(oldSettings);
-        });
-
-        test('accepts new settings', async() => {
-          const oldSettings: Settings = JSON.parse((await rdctl(['list-settings'])).stdout);
-          const body: RecursivePartial<Settings> = {
-            ...(os.platform() === 'win32' ? {} : {
-              virtualMachine: {
-                memoryInGB: oldSettings.virtualMachine.memoryInGB + 1,
-                numberCPUs: oldSettings.virtualMachine.numberCPUs + 1,
-              },
-            }),
-            version:     CURRENT_SETTINGS_VERSION,
-            application: {
-              // XXX: Can't change adminAccess until we can process the sudo-request dialog (and decline it)
-              // adminAccess: !oldSettings.application.adminAccess,
-              telemetry: { enabled: !oldSettings.application.telemetry.enabled },
-              updater:   { enabled: !oldSettings.application.updater.enabled },
-              debug:     !oldSettings.application.debug,
-            },
-            // This field is to force a restart
-            kubernetes: { port: oldSettings.kubernetes.port + 1 },
-          };
-
-          if (process.platform !== 'win32' && body.application !== undefined) {
-            body.application.pathManagementStrategy = getAlternateSetting(oldSettings,
-              'application.pathManagementStrategy',
-              PathManagementStrategy.Manual,
-              PathManagementStrategy.RcFiles);
-          }
-          const { stdout, stderr, error } = await rdctl(['api', '/v1/settings', '-X', 'PUT', '-b', JSON.stringify(body)]);
-
-          expect({
-            stdout, stderr, error,
-          }).toEqual({
-            stdout: expect.stringContaining('reconfiguring Rancher Desktop to apply changes'),
-            stderr: '',
-            error:  undefined,
-          });
-          const newSettings: Settings = JSON.parse((await rdctl(['list-settings'])).stdout);
-
-          expect(newSettings).toEqual(_.merge(oldSettings, body));
-
-          // And now reinstate the old prefs so other tests that count on them will pass.
-          const result = await rdctl(['api', '/v1/settings', '-X', 'PUT', '-b', JSON.stringify(oldSettings)]);
-
-          expect(result.stderr).toEqual('');
-          // Have to do this because we don't have any other way to see the current missing progress bar
-          // and have the next  `progressBecomesReady` test pass prematurely.
-
-          // Wait until progress bar show up. It takes roughly ~60s to start in CI
-          const progressBar = page.locator('.progress');
-
-          await waitForRestartVM(progressBar);
-
-          // Since we just applied new settings, we must wait for the backend to restart.
-          while (await progressBar.count() > 0) {
-            await progressBar.waitFor({ state: 'detached', timeout: Math.round(240_000) });
-          }
-        });
-      });
-
       test('complains about options not intended for current platform', async() => {
         // playwright doesn't support test.each
         // See https://github.com/microsoft/playwright/issues/7036 for the discussion
@@ -1414,37 +1295,6 @@ test.describe('Command server', () => {
           await fs.promises.rm(tmpDir, { recursive: true, force: true });
         }
       });
-    });
-    test('should verify nerdctl can talk to containerd', async() => {
-      const { stdout } = await rdctl(['list-settings']);
-      const settings: Settings = JSON.parse(stdout);
-
-      if (settings.containerEngine.name !== ContainerEngine.CONTAINERD) {
-        const payloadObject: RecursivePartial<Settings> = {
-          version:         CURRENT_SETTINGS_VERSION,
-          containerEngine: { name: ContainerEngine.CONTAINERD },
-        };
-        const navPage = new NavPage(page);
-
-        await tool('rdctl', 'api', '/v1/settings', '--method', 'PUT', '--body', JSON.stringify(payloadObject));
-        await waitForRestartVM(page.locator('.progress'));
-        await navPage.progressBecomesReady();
-      }
-      const output = await retry(() => tool('nerdctl', 'info'));
-
-      expect(output).toMatch(/Server Version:\s+v?[.0-9]+/);
-    });
-    test('should verify docker can talk to dockerd', async() => {
-      const navPage = new NavPage(page);
-
-      await tool('rdctl', 'set', '--container-engine', 'moby');
-      await expect(navPage.progressBar).not.toBeHidden();
-      await waitForRestartVM(navPage.progressBar);
-      await navPage.progressBecomesReady();
-      await expect(navPage.progressBar).toBeHidden();
-      const output = await retry(() => tool('docker', 'info'), { delay: 500, tries: 60 });
-
-      expect(output).toMatch(/Server Version:\s+v?[.0-9]+/);
     });
   });
